Updated patient landing buttons and organized prep files
Added a button that will lead to the patient details page from the patient cards on the patient info landing page.

Also, added the mockup files to the prepFiles folder and made a new folder for the SQL files. The project will need to be zipped and submitted to the class. So all relevant planning files need to be in the project. Additionally, the planning doc for the whole project has been added to the folder as well.